+++ src/components/pages/head-guest-page.tsx
