+++ src/components/pages/movie-details-page.tsx
