+++ src/components/pages/player-page.tsx
