+++ src/components/pages/player-pause-page.tsx
