+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import {App} from './App';
+import {App} from './app';
 import {Film} from './types/film';
 
 const root = ReactDOM.createRoot(
